Extract Fragment for scan view
